qt: (refactor) split "send tab" out from main_window.py
