+++ electrum/gui/qt/transaction_dialog.py
@@ -242,7 +242,7 @@ class BaseTxDialog(QDialog, MessageBoxMixin):
 
     def do_broadcast(self):
         self.main_window.push_top_level_window(self)
-        self.main_window.save_pending_invoice()
+        self.main_window.send_tab.save_pending_invoice()
         try:
             self.main_window.broadcast_transaction(self.tx)
         finally:
